requirements.txt
python-dotenv

pandas
polars
pyarrow
requests

apache-airflow

beautifulsoup4
selenium
webdriver-manager


pytesseract

streamlit
